feature: results in CSV, new methods: refresh_csv(), get_results_csv(), refresh_into_dataframe()
This makes use of the Dune API with execution results in CSV format

This is ideal for loading the data into Pandas, which is a common usecase.

Additionally, it is also useful for large results, because it uses less CPU and memory:
 - it avoids json parsing the (possibly large) result
 - doesn't create python dictionaries for every single row
 - doesn't duplicate in memory the column names, which would be present
   in the row dictionaries as keys

Please note that the result is still fully read into memory, until DuneAPI provides
pagination support, this is unavoidable.