PRO-211: Add support for is_private in upload_csv
As per https://github.com/duneanalytics/docs/pull/399 , the Dune Upload
API now supports uploading private datasets. With this commit, we add
this functionality to the upload_csv function.

I recognize that this functionality gets inside the "Plus Subscription"
section even though it is premium. I decided that this was the best
option and cleaner than creating a `upload_csv_private` function.

Closes #97.