Implement get_latest_result_dataframe helper function
The get_latest_result implementation parses json,
which is slow for large results. The csv endpoint
is a lot more efficient, especially if you just
need to put the data into a pandas DataFrame.
get_latest_result_dataframe is a helper function
that calls download_csv and returns a DataFrame
from the result.